feat: add delete logic depend of FK, add delete all not active players, add validations for methods post and update

+++ src/main/java/com/example/squid_game_api/services/GameService.java
@@ -4,6 +4,8 @@ import com.example.squid_game_api.dto.GameRequest;
 import com.example.squid_game_api.dto.GameResponse;
 import com.example.squid_game_api.entities.Game;
 import com.example.squid_game_api.repositories.GameRepository;
+import com.example.squid_game_api.repositories.ParticipationRepository;
+import jakarta.transaction.Transactional;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalTime;
@@ -14,9 +16,11 @@ import java.util.Optional;
 public class GameService {
 
     private final GameRepository gameRepository;
+    private final ParticipationRepository participationRepository;
 
-    public GameService(GameRepository gameRepository) {
+    public GameService(GameRepository gameRepository, ParticipationRepository participationRepository) {
         this.gameRepository = gameRepository;
+        this.participationRepository = participationRepository;
     }
 
     // Create a new game
@@ -65,13 +69,21 @@ public class GameService {
     }
 
     // Delete a game
+    @Transactional
     public void deleteGameById(Long id) {
         if (!gameRepository.existsById(id)) {
             throw new RuntimeException("Game with ID " + id + " not found.");
         }
+
+        boolean hasParticipations = participationRepository.existsByGameId(id);
+        if (hasParticipations) {
+            throw new RuntimeException("Game with ID " + id + " cannot be deleted because it has associated participations.");
+        }
+
         gameRepository.deleteById(id);
     }
-    //Check if game is already has been finish
+
+    //Check if game is already has been finished
     public boolean isGameFinished(Game game) {
         LocalTime now = LocalTime.now();
         return now.isAfter(game.getEndTime());
